+++ ruoyi-common/src/main/java/com/ruoyi/common/enums/DataSourceType.java
@@ -1,0 +1,19 @@
+package com.ruoyi.common.enums;
+
+/**
+ * 数据源
+ *
+ * @author ruoyi
+ */
+public enum DataSourceType
+{
+    /**
+     * 主库
+     */
+    MASTER,
+
+    /**
+     * 从库
+     */
+    SLAVE
+}